# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi Main API",
        "description": "Tulang punggung dari berbagai aplikasi.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorinci/available-colors": {
            "get": {
                "tags": [
                    "Motorinci Available Colors"
                ],
                "summary": "Mendapatkan daftar warna yang tersedia untuk motor",
                "operationId": "getMotorinciAvailableColorsList",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Available Colors"
                ],
                "summary": "Menambahkan warna baru untuk sebuah motor",
                "operationId": "storeMotorinciAvailableColor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "motor_id",
                                    "color_id"
                                ],
                                "properties": {
                                    "motor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "color_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "Gambar motor dengan warna terkait",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201_Created"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    }
                }
            }
        },
        "/api/motorinci/available-colors/{id}": {
            "get": {
                "tags": [
                    "Motorinci Available Colors"
                ],
                "summary": "Mendapatkan detail warna tersedia berdasarkan ID",
                "operationId": "getMotorinciAvailableColorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Available Color ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Available Colors"
                ],
                "summary": "Update data warna tersedia",
                "description": "Gunakan method POST dengan _method=PUT di dalam form-data untuk request PUT/PATCH.",
                "operationId": "updateMotorinciAvailableColor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Available Color ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "motor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "color_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "image": {
                                        "description": "Gambar baru (opsional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorinci Available Colors"
                ],
                "summary": "Menghapus data warna tersedia",
                "operationId": "deleteMotorinciAvailableColor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Available Color ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successful"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operasi berhasil"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan"
                    }
                },
                "type": "object"
            },
            "ValidationErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data yang diberikan tidak valid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "Pesan error."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200_Success": {
                "description": "Operasi berhasil dieksekusi.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "201_Created": {
                "description": "Sumber daya berhasil dibuat.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
...
            "401_Unauthorized": {
                "description": "Autentikasi gagal. Token tidak valid atau tidak ada.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
...
            "404_NotFound": {
                "description": "Sumber daya yang diminta tidak ditemukan.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "422_UnprocessableContent": {
                "description": "Validasi gagal.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorJsonResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Bearer Anda. Contoh: 'Bearer 1|abcdef...'",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}